# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Prospero API",
    "description": "",
    "license": {
      "name": ""
    },
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/books/{slug}/resources": {
      "get": {
        "tags": [
          "resources"
        ],
        "summary": "List resources for a sentence range (public)",
        "operationId": "list_resources",
        "parameters": [
          {
            "name": "slug",
            "in": "path",
            "description": "Book slug",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "start",
            "in": "query",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "end",
            "in": "query",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "kind",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Resources for the sentence range",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResourceListResponse"
                }
              }
            }
          },
          "404": {
            "description": "Book not found"
          }
        }
      },
      "post": {
        "tags": [
          "resources"
        ],
        "summary": "Create a new resource (editor only)",
        "operationId": "create_resource",
        "parameters": [
          {
            "name": "slug",
            "in": "path",
            "description": "Book slug",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateResourceRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Resource created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResourceListResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid input"
          },
          "401": {
            "description": "Not authenticated"
          },
          "403": {
            "description": "Insufficient permissions"
          }
        }
      }
    },
    "/api/books/{slug}/resources/{id}": {
      "put": {
        "tags": [
          "resources"
        ],
        "summary": "Update an existing resource (editor only)",
        "operationId": "update_resource",
        "parameters": [
          {
            "name": "slug",
            "in": "path",
            "description": "Book slug",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "id",
            "in": "path",
            "description": "Resource ID",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateResourceRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Resource updated"
          },
          "400": {
            "description": "Invalid input"
          },
          "401": {
            "description": "Not authenticated"
          },
          "403": {
            "description": "Insufficient permissions"
          },
          "404": {
            "description": "Resource not found"
          }
        }
      },
      "delete": {
        "tags": [
          "resources"
        ],
        "summary": "Soft-delete a resource (editor only)",
        "operationId": "delete_resource",
        "parameters": [
          {
            "name": "slug",
            "in": "path",
            "description": "Book slug",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "id",
            "in": "path",
            "description": "Resource ID",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Resource archived"
          },
          "401": {
            "description": "Not authenticated"
          },
          "403": {
            "description": "Insufficient permissions"
          },
          "404": {
            "description": "Resource not found"
          }
        }
      }
    },
...
    "/api/persons": {
      "get": {
        "tags": [
          "persons"
        ],
        "summary": "Search persons by name",
        "operationId": "search_persons",
        "parameters": [
          {
            "name": "q",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of matching persons",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/PersonResponse"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Not authenticated"
          },
          "403": {
            "description": "Insufficient permissions"
          }
        }
      },
      "post": {
        "tags": [
          "persons"
        ],
        "summary": "Create a new person",
        "operationId": "create_person",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreatePersonRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Person created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PersonResponse"
                }
              }
            }
          },
          "401": {
            "description": "Not authenticated"
          },
          "403": {
            "description": "Insufficient permissions"
          }
        }
      }
    },
    "/api/persons/{id}": {
      "put": {
        "tags": [
          "persons"
        ],
        "summary": "Update an existing person",
        "operationId": "update_person",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Person ID",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdatePersonRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Person updated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PersonResponse"
                }
              }
            }
          },
          "401": {
            "description": "Not authenticated"
          },
          "403": {
            "description": "Insufficient permissions"
          },
          "404": {
            "description": "Person not found"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CreatePersonRequest": {
        "type": "object",
        "required": [
          "name"
        ],
        "properties": {
          "name": {
            "type": "string"
          },
          "sort_name": {
            "type": [
              "string",
              "null"
            ]
          }
        }
      },
      "CreateResourceRequest": {
        "type": "object",
        "required": [
          "resource_type",
          "sentence_start",
          "sentence_kind"
        ],
        "properties": {
          "admin_notes": {
            "type": [
              "string",
              "null"
            ]
          },
          "editor_note": {
            "type": [
              "string",
              "null"
            ]
          },
          "is_featured": {
            "type": [
              "boolean",
              "null"
            ]
          },
          "quoted_text": {
            "type": [
              "string",
              "null"
            ]
          },
          "resource_type": {
            "type": "string"
          },
          "sentence_end": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int32"
          },
          "sentence_kind": {
            "type": "string"
          },
          "sentence_start": {
            "type": "integer",
            "format": "int32"
          },
          "source_id": {
            "type": [
              "string",
              "null"
            ]
          },
          "source_location_freeform": {
            "type": [
              "string",
              "null"
            ]
          },
          "source_page_end": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int32"
          },
          "source_page_start": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int32"
          },
          "verbatim_kind": {
            "type": [
              "string",
              "null"
            ]
          }
        }
      },
...
      "ParentSourceResponse": {
        "type": "object",
        "required": [
          "id",
          "source_type",
          "title",
          "persons"
        ],
        "properties": {
          "id": {
            "type": "string"
          },
          "persons": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/SourcePersonResponse"
            }
          },
          "publication_year": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int32"
          },
          "publisher": {
            "type": [
              "string",
              "null"
            ]
          },
          "source_type": {
            "type": "string"
          },
          "title": {
            "type": "string"
          }
        }
      },
      "PersonResponse": {
        "type": "object",
        "required": [
          "id",
          "name"
        ],
        "properties": {
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "sort_name": {
            "type": [
              "string",
              "null"
            ]
          }
        }
      },
...
      "ResourceListResponse": {
        "type": "object",
        "required": [
          "resources"
        ],
        "properties": {
          "resources": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ResourceResponse"
            }
          }
        }
      },
      "ResourceResponse": {
        "type": "object",
        "required": [
          "id",
          "resource_type",
          "anchor_sentence_start_number",
          "sentence_kind",
          "is_featured",
          "created_at"
        ],
        "properties": {
          "admin_notes": {
            "type": [
              "string",
              "null"
            ]
          },
          "anchor_sentence_end_number": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int32"
          },
          "anchor_sentence_start_number": {
            "type": "integer",
            "format": "int32"
          },
          "created_at": {
            "type": "string"
          },
          "editor_note": {
            "type": [
              "string",
              "null"
            ]
          },
          "id": {
            "type": "string"
          },
          "is_featured": {
            "type": "boolean"
          },
          "quoted_text": {
            "type": [
              "string",
              "null"
            ]
          },
          "resource_type": {
            "type": "string"
          },
          "sentence_kind": {
            "type": "string"
          },
          "source": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/SourceResponse"
              }
            ]
          },
          "source_location_freeform": {
            "type": [
              "string",
              "null"
            ]
          },
          "source_page_end": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int32"
          },
          "source_page_start": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int32"
          },
          "verbatim_kind": {
            "type": [
              "string",
              "null"
            ]
          }
        }
      },
...
      "SourcePersonResponse": {
        "type": "object",
        "required": [
          "person_id",
          "name",
          "role",
          "position"
        ],
        "properties": {
          "name": {
            "type": "string"
          },
          "person_id": {
            "type": "string"
          },
          "position": {
            "type": "integer",
            "format": "int32"
          },
          "role": {
            "type": "string"
          },
          "sort_name": {
            "type": [
              "string",
              "null"
            ]
          }
        }
      },
      "SourceResponse": {
        "type": "object",
        "required": [
          "id",
          "source_type",
          "title",
          "persons"
        ],
        "properties": {
          "doi": {
            "type": [
              "string",
              "null"
            ]
          },
          "edition": {
            "type": [
              "string",
              "null"
            ]
          },
          "id": {
            "type": "string"
          },
          "isbn": {
            "type": [
              "array",
              "null"
            ],
            "items": {
              "type": "string"
            }
          },
          "journal_name": {
            "type": [
              "string",
              "null"
            ]
          },
          "page_end": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int32"
          },
          "page_start": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int32"
          },
          "parent": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/ParentSourceResponse"
              }
            ]
          },
          "persons": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/SourcePersonResponse"
            }
          },
          "publication_year": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int32"
          },
          "publisher": {
            "type": [
              "string",
              "null"
            ]
          },
          "source_type": {
            "type": "string"
          },
          "title": {
            "type": "string"
          },
          "url": {
            "type": [
              "string",
              "null"
            ]
          },
          "volume": {
            "type": [
              "string",
              "null"
            ]
          }
        }
      },
...
      "UpdatePersonRequest": {
        "type": "object",
        "properties": {
          "name": {
            "type": [
              "string",
              "null"
            ]
          },
          "sort_name": {
            "type": [
              "string",
              "null"
            ]
          }
        }
      },
...
      "UpdateResourceRequest": {
        "type": "object",
        "properties": {
          "admin_notes": {
            "type": [
              "string",
              "null"
            ]
          },
          "editor_note": {
            "type": [
              "string",
              "null"
            ]
          },
          "is_featured": {
            "type": [
              "boolean",
              "null"
            ]
          },
          "quoted_text": {
            "type": [
              "string",
              "null"
            ]
          },
          "resource_type": {
            "type": [
              "string",
              "null"
            ]
          },
          "sentence_end": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int32"
          },
          "sentence_kind": {
            "type": [
              "string",
              "null"
            ]
          },
          "sentence_start": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int32"
          },
          "source_id": {
            "type": [
              "string",
              "null"
            ]
          },
          "source_location_freeform": {
            "type": [
              "string",
              "null"
            ]
          },
          "source_page_end": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int32"
          },
          "source_page_start": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int32"
          },
          "verbatim_kind": {
            "type": [
              "string",
              "null"
            ]
          }
        }
      },
...
    }
  }
}